Refactor: merge 2 stimulus controllers together
Merge `progress-preview` and `touched-field-marker` stimulus controllers
together into `preview` controller.

They are sharing some state, the touched status, showing that they need
to be grouped together. It will make it easier to deal with the touched
state later.
