+++ app/components/work_packages/progress/status_based/modal_body_component.html.erb
@@ -6,10 +6,9 @@
         id: "progress-form",
         html: { autocomplete: "off" },
         data: { "application-target": "dynamic",
-                "work-packages--progress--preview-progress-target": "form",
+                "work-packages--progress--preview-target": "form",
                 controller: "work-packages--progress--focus-field " \
-                  "work-packages--progress--preview-progress " \
-                  "work-packages--progress--touched-field-marker" }
+                            "work-packages--progress--preview" }
       ) do |f| %>
     <%= flex_layout do |modal_body| %>
       <% modal_body.with_row(classes: "FormControl-horizontalGroup--sm-vertical") do |_fields| %>
