+++ app/forms/work_packages/pre_14_4_progress_form.rb
@@ -91,11 +91,11 @@ class WorkPackages::Pre144ProgressForm < ApplicationForm
 
         group.hidden(name: :status_id_touched,
                      value: @touched_field_map["status_id_touched"] || false,
-                     data: { "work-packages--progress--touched-field-marker-target": "touchedFieldInput",
+                     data: { "work-packages--progress--preview-target": "touchedFieldInput",
                              "referrer-field": "work_package[status_id]" })
         group.hidden(name: :estimated_hours_touched,
                      value: @touched_field_map["estimated_hours_touched"] || false,
-                     data: { "work-packages--progress--touched-field-marker-target": "touchedFieldInput",
+                     data: { "work-packages--progress--preview-target": "touchedFieldInput",
                              "referrer-field": "work_package[estimated_hours]" })
       else
         render_text_field(group, name: :estimated_hours, label: I18n.t(:label_work))
@@ -105,11 +105,11 @@ class WorkPackages::Pre144ProgressForm < ApplicationForm
 
         group.hidden(name: :estimated_hours_touched,
                      value: @touched_field_map["estimated_hours_touched"] || false,
-                     data: { "work-packages--progress--touched-field-marker-target": "touchedFieldInput",
+                     data: { "work-packages--progress--preview-target": "touchedFieldInput",
                              "referrer-field": "work_package[estimated_hours]" })
         group.hidden(name: :remaining_hours_touched,
                      value: @touched_field_map["remaining_hours_touched"] || false,
-                     data: { "work-packages--progress--touched-field-marker-target": "touchedFieldInput",
+                     data: { "work-packages--progress--preview-target": "touchedFieldInput",
                              "referrer-field": "work_package[remaining_hours]" })
       end
       group.fields_for(:initial) do |builder|
@@ -195,9 +195,8 @@ class WorkPackages::Pre144ProgressForm < ApplicationForm
   end
 
   def default_field_options(name)
-    data = { "work-packages--progress--preview-progress-target": "progressInput",
-             "work-packages--progress--touched-field-marker-target": "progressInput",
-             action: "input->work-packages--progress--touched-field-marker#markFieldAsTouched" }
+    data = { "work-packages--progress--preview-target": "progressInput",
+             action: "input->work-packages--progress--preview#markFieldAsTouched" }
 
     if @focused_field == name
       data[:"work-packages--progress--focus-field-target"] = "fieldToFocus"
