+++ app/forms/work_packages/progress_form/initial_values_form.rb
@@ -57,7 +57,7 @@ class WorkPackages::ProgressForm
     def hidden_initial_field(form, name:)
       form.hidden(name:,
                   value: work_package.public_send(:"#{name}_was"),
-                  data: { "work-packages--progress--touched-field-marker-target": "initialValueInput",
+                  data: { "work-packages--progress--preview-target": "initialValueInput",
                           "referrer-field": "work_package[#{name}]" })
     end
   end
